Chore: Redis 설정 변경
