+++ src/main/java/com/project/goalchallenge/global/config/redis/RedisConfig.java
@@ -1,0 +1,63 @@
+package com.project.goalchallenge.global.config.redis;
+
+import java.util.HashMap;
+import java.util.Map;
+import lombok.RequiredArgsConstructor;
+import org.redisson.Redisson;
+import org.redisson.api.RedissonClient;
+import org.redisson.codec.JsonJacksonCodec;
+import org.redisson.config.Config;
+import org.redisson.spring.cache.CacheConfig;
+import org.redisson.spring.cache.RedissonSpringCacheManager;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.data.redis.RedisProperties;
+import org.springframework.cache.annotation.EnableCaching;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
+import org.springframework.data.redis.serializer.StringRedisSerializer;
+
+@Configuration
+@EnableCaching
+@EnableRedisRepositories
+@RequiredArgsConstructor
+public class RedisConfig {
+
+  private final RedisProperties redisProperties;
+
+  @Bean
+  public RedissonClient redissonClient() {
+    Config config = new Config();
+    config.useSingleServer()
+        .setAddress("redis://" + redisProperties.getHost() + ":" + redisProperties.getPort());
+    config.setCodec(new JsonJacksonCodec());
+    return Redisson.create(config);
+  }
+
+  @Bean
+  public RedissonSpringCacheManager cacheManager(RedissonClient redissonClient) {
+    Map<String, CacheConfig> configMap = new HashMap<>();
+
+    configMap.put("challengeParticipantsDuplicateCache", new CacheConfig(60 * 60 * 1000, 10 * 60 * 1000));
+    return new RedissonSpringCacheManager(redissonClient, configMap);
+  }
+
+
+  @Bean
+  public RedisConnectionFactory redisConnectionFactory() {
+    return new LettuceConnectionFactory(redisProperties.getHost(), redisProperties.getPort());
+  }
+
+  @Bean
+  public RedisTemplate<String, Object> redisTemplate() {
+    RedisTemplate<String, Object> redisTemplate = new RedisTemplate<>();
+    redisTemplate.setConnectionFactory(redisConnectionFactory());
+    redisTemplate.setKeySerializer(new StringRedisSerializer());
+    redisTemplate.setValueSerializer(new StringRedisSerializer());
+    return redisTemplate;
+  }
+
+}
